[token-client-spring-demo]: update after refactoring of token-client-spring module
