+++ token-client-spring-demo/src/main/java/no/nav/security/token/support/demo/spring/config/ApplicationConfiguration.java
@@ -1,6 +1,6 @@
 package no.nav.security.token.support.demo.spring.config;
 
-import no.nav.security.token.support.oauth2.EnableOAuth2Client;
+import no.nav.security.token.support.client.spring.oauth2.EnableOAuth2Client;
 import no.nav.security.token.support.spring.api.EnableJwtTokenValidation;
 import org.springframework.context.annotation.Configuration;
